different headers for logged/unlogged
